+++ src/main/java/fr/insee/genesis/domain/model/surveyunit/rawdata/RawResponseModel.java
@@ -1,7 +1,6 @@
 package fr.insee.genesis.domain.model.surveyunit.rawdata;
 
 import fr.insee.genesis.domain.model.surveyunit.Mode;
-import lombok.Builder;
 import org.bson.types.ObjectId;
 
 import java.time.LocalDateTime;